backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.0.3
snowflake-connector-python==3.6.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-anthropic>=0.1.15
anthropic>=0.16.0
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1